[Update] Added basic ChatroomCreateFragment boilerplate + delegate interface abstractions
