--- app/src/main/java/com/example/hobbyfi/ui/create/ChatroomCreateFragment.kt
@@ -1,21 +1,0 @@
-package com.example.hobbyfi.ui.create
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.hobbyfi.R
-import com.example.hobbyfi.ui.base.BaseFragment
-
-class ChatroomCreateFragment : BaseFragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_chatroom_create, container, false)
-    }
-
-}
