+++ app/src/main/java/com/example/hobbyfi/viewmodels/base/NameDescriptionBindable.kt
@@ -1,0 +1,4 @@
+package com.example.hobbyfi.viewmodels.base
+
+interface NameDescriptionBindable {
+}
