+++ app/src/main/java/com/example/hobbyfi/viewmodels/base/NameDescriptionBindableViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.hobbyfi.viewmodels.base
+
+class NameDescriptionBindableViewModel {
+}
