+++ app/src/main/java/com/example/hobbyfi/viewmodels/factories/AuthChatroomViewModelFactory.kt
@@ -1,0 +1,4 @@
+package com.example.hobbyfi.viewmodels.factories
+
+class AuthChatroomViewModelFactory {
+}
